cleanup implementation for 1.0.0

+++ jarowinkler/scorer.hpp
@@ -7,7 +7,7 @@ static inline double jaro_similarity_func(const RF_String& s1, const RF_String& 
         return jaro_winkler::jaro_similarity(first1, last1, first2, last2, score_cutoff);
     });
 }
-static inline bool JaroSimilarityInit(RF_ScorerFunc* self, const RF_Kwargs*, size_t str_count, const RF_String* str)
+static inline bool JaroSimilarityInit(RF_ScorerFunc* self, const RF_Kwargs*, int64_t str_count, const RF_String* str)
 {
     return scorer_init_f64<jaro_winkler::CachedJaroSimilarity>(self, str_count, str);
 }
@@ -19,7 +19,7 @@ static inline double jaro_winkler_similarity_func(const RF_String& s1, const RF_
         return jaro_winkler::jaro_winkler_similarity(first1, last1, first2, last2, prefix_weight, score_cutoff);
     });
 }
-static inline bool JaroWinklerSimilarityInit(RF_ScorerFunc* self, const RF_Kwargs* kwargs, size_t str_count, const RF_String* str)
+static inline bool JaroWinklerSimilarityInit(RF_ScorerFunc* self, const RF_Kwargs* kwargs, int64_t str_count, const RF_String* str)
 {
     return scorer_init_f64<jaro_winkler::CachedJaroWinklerSimilarity>(self, str_count, str, *(double*)(kwargs->context));
 }
